requirements.txt
fastapi
pydantic
uvicorn
tensorflow
numpy
Pillow
python-multipart